\name{GUIP1-package}

\alias{GUIP1-package}

\docType{package}

\title{
Graphical User Interface for Phase I Designs
}

\description{
This package provides graphical user interfaces for several model-guided adaptive designs (CRM, TITE-CRM, EWOC, TITE-EWOC).
}

\details{
This package contains 1 main function called GUIP1 allowing users to run 5 model-guided adaptive designs:
-CRMB, continual reassessment method with bayes' estimator,
-CRML, continual reassessment method with mle,
-TITECRM, TIme-To-Event continual reassessment method,
-EWOC, Escalation With Overdose Control,
-TEWOC, TIme-To-Event Escalation With Overdose Control.
}


\seealso{
\link[dfcrm:crm]{crm}
\link[bcrm:bcrm]{ewoc} 
\link[dfcrm:titecrm]{titecrm}
}
