\name{GUIP1}

\alias{GUIP1}

\title{
Graphical User Interface for Phase I designs (GUIP1)
}

\description{
This function allows to launch a graphical user interface for several model-guided adaptive designs (CRM, TITE-CRM, EWOC, TITE-EWOC).
}

\usage{
GUIP1()
}

\details{
Allowing users to run 5 model-guided adaptive designs:
-CRMB, continual reassessment method with bayes' estimator,
-CRML, continual reassessment method with mle,
-TITECRM, TIme-To-Event continual reassessment method,
-EWOC, Escalation With Overdose Control,
-TEWOC, TIme-To-Event Escalation With Overdose Control.
}

\value{
Not applicable
}

\references{
O'Quigley, J., Pepe, M., and Fisher, L. (1990). Continual reassessment method: a practical design for phase 1 clinical trials in cancer. Biometrics 46, 33-48.

O'Quigley, J., and Shen, L.Z. (1996). Continual reassessment method: a likelihood approach. Biometrics 52, 673-684.

Babb, J., Rogatko, A., and Zacks, S. (1998). Cancer phase I clinical trials: efficient dose escalation with overdose control. Stat. Med. 17, 1103-1120.

Cheung, Y.K., and Chappell, R. (2000). Sequential designs for phase I clinical trials with late-onset toxicities. Biometrics 56, 1177-1182.

Mauguen, A., Le Deley, M.C., and Zohar, S. (2011). Dose-finding approach for dose escalation with overdose control considering incomplete observations. Stat. Med. 30, 1584-1594.
}
